Completed maze Generator
the completed maze generator has been added, next step make it sort of into a game, with difficulties. As well we can mark the beggining location and have a player to control too go through the maze leaving behind a green line in the center of where we have already traveled and if we make it to the end of the maze the game resets to a new maze or something with a bigger size.